+++ packages/testing/src/mockOperationContext.ts
@@ -1,37 +1,5 @@
-import { Operation, OperationArgs } from '@fleur/fleur'
-import { ActionIdentifier } from '@fleur/fleur'
 import { MockStore } from './mockStore'
 import { MockContextBase } from './MockContextBase'
-
-export class MockOperationContext extends MockContextBase {
-  public dispatches: { action: ActionIdentifier<any>; payload: any }[] = []
-
-  public executeOperation = async <O extends Operation>(
-    operation: O,
-    ...args: OperationArgs<O>
-  ): Promise<void> => {
-    await Promise.resolve(operation(this as any, ...args))
-  }
-
-  public dispatch = <AI extends ActionIdentifier<any>>(
-    action: AI,
-    payload: ReturnType<AI>,
-  ): void => {
-    this.dispatches.push({ action, payload })
-    this.mockStores.forEach(({ store }) => {
-      Object.keys(store)
-        .filter(
-          key =>
-            (store as any)[key] != null && (store as any)[key].__fleurHandler,
-        )
-        .forEach(key => {
-          if ((store as any)[key].__action === action) {
-            ;(store as any)[key].producer(payload)
-          }
-        })
-    })
-  }
-}
 
 export const mockOperationContext = ({
   stores,
@@ -39,6 +7,6 @@ export const mockOperationContext = ({
 }: {
   stores: readonly MockStore[]
   mocks: Map<any, any>
-}): MockOperationContext => {
-  return new MockOperationContext({ stores, mocks })
+}): MockContextBase => {
+  return new MockContextBase({ stores, mocks })
 }