+++ packages/create-next-app/CHANGELOG.md
@@ -1,3 +1,11 @@
+# @fleur/create-next-app Changelog
+
+### 0.7.0
+
+- Change package name to `@fleur/create-next-app`
+- Support Next.js 10
+- Follow @fleur/next update
+
 ### 0.6.0
 - dfc48c4 Fix deprecated type using